add test for remove function update readme file
